on { Add(val)=>{}; Reset=>{} } → enum TaskMsg + 变体 send 改写;
on { "ping"=>{} } → match msg.as_str()。007_named_variants 端到端通过。
parser 注册 task 变体名(解决 h.send(Reset) undefined variable)。

+++ crates/auto-lang/test/a2r/22_actors/006_state_increment/state_increment.expected.rs
@@ -7,7 +7,7 @@ use a2r_std::*;
 
 #[derive(Clone, Debug)]
 struct Counter {
-    count: i32,
+    count: i64,
 }
 
 impl Counter {
